Add. 6.0 Twitter Cloning 학습 - Cleaning JS
