+++ 3. 트위터/jwitter/src/components/AuthForm.jsx
@@ -1,0 +1,64 @@
+import React, { useState } from "react";
+import { createUserWithEmailAndPassword, getAuth, signInWithEmailAndPassword } from "firebase/auth";
+
+const AuthForm = () => {
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+  const [newAcount, setNewAccount] = useState(true);
+  const [error, setError] = useState('');
+  const onChange = (e) => {
+    const { target: { name, value } } = e;
+    if (name === 'email') {
+      setEmail(value);
+    } else if (name === 'password') {
+      setPassword(value);
+    }
+  }
+  const onSubmit = async (e) => {
+    e.preventDefault();
+    try {
+      let data;
+      const auth = getAuth();
+      if (newAcount) {
+        // create 
+        data = await createUserWithEmailAndPassword(auth, email, password);
+      } else {
+        // log in
+        data = await signInWithEmailAndPassword(auth, email, password);
+      }
+    } catch (error) {
+      setError(error.message);
+    }
+  };
+
+  const toggleAccount = () => setNewAccount(prev => !prev);
+  return (
+    <>
+      <form onSubmit={onSubmit}>
+        <input
+          name='email'
+          type="text"
+          placeholder='Email'
+          required
+          value={email}
+          onChange={onChange}
+        />
+        <input
+          name='password'
+          type="password"
+          placeholder='Password'
+          required
+          value={password}
+          onChange={onChange}
+        />
+        <input type="submit" value={newAcount ? "Create Account" : "Log In"} />
+        {error}
+      </form>
+      <span onClick={toggleAccount}>
+        {newAcount ? 'Sign In' : 'Create Account'}
+      </span>
+    </>
+  )
+}
+
+export default AuthForm;
