+++ 3. 트위터/jwitter/src/components/Navigation.jsx
@@ -9,7 +9,9 @@ const Navigation = ({userObj}) => {
           <Link to="/">Home</Link>
         </li>
         <li>
-          <Link to="/profile">{ userObj.displayName}의 Profile</Link>
+          <Link to="/profile">{userObj?.displayName?.length
+? `${userObj.displayName}'s Profile`
+: "Profile"}</Link>
         </li>
       </ul>
     </nav>
