+++ 3. 트위터/jwitter/src/routes/Auth.js
@@ -1,48 +1,15 @@
-import React, { useState } from 'react';
+import React from 'react';
 import {
-  createUserWithEmailAndPassword,
-  getAuth,
   GithubAuthProvider,
   GoogleAuthProvider,
-  signInWithEmailAndPassword,
   signInWithPopup
 } from 'firebase/auth';
 import { auth } from 'fbase';
+import AuthForm from 'components/AuthForm';
 
 const Auth = () => {
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [newAcount, setNewAccount] = useState(true);
-  const [error, setError] = useState('');
-  const onChange = (e) => {
-    const { target: { name, value } } = e;
-    if (name === 'email') {
-      setEmail(value);
-    } else if (name === 'password') {
-      setPassword(value);
-    }
-  }
-  const onSubmit = async (e) => {
-    e.preventDefault();
-    try {
-      let data;
-      const auth = getAuth();
-      if (newAcount) {
-        // create 
-        data = await createUserWithEmailAndPassword(auth, email, password);
-      } else {
-        // log in
-        data = await signInWithEmailAndPassword(auth, email, password);
-      }
-      console.log(data);
-    } catch (error) {
-      setError(error.message);
-    }
-  };
-
-  const toggleAccount = () => setNewAccount(prev => !prev);
   const onSocialClick = async (e) => {
-    const { target: { name, value } } = e;
+    const { target: { name } } = e;
     let provider;
     if (name === 'google') {
       provider = new GoogleAuthProvider();
@@ -55,29 +22,7 @@ const Auth = () => {
 
   return (
     <>
-      <form onSubmit={onSubmit}>
-        <input
-          name='email'
-          type="text"
-          placeholder='Email'
-          required
-          value={email}
-          onChange={onChange}
-        />
-        <input
-          name='password'
-          type="password"
-          placeholder='Password'
-          required
-          value={password}
-          onChange={onChange}
-        />
-        <input type="submit" value={newAcount ? "Create Account" : "Log In"} />
-        {error}
-      </form>
-      <span onClick={toggleAccount}>
-        {newAcount ? 'Sign In' : 'Create Account'}
-      </span>
+      <AuthForm />
       <div>
         <button onClick={onSocialClick} name='google'>Continue with Google</button>
         <button onClick={onSocialClick} name='github'>Continue with Github</button>
